refactor cards page
