+++ frontend/package.json
@@ -35,7 +35,7 @@
     "cmdk": "^0.2.0",
     "copy-to-clipboard": "^3.3.3",
     "jotai": "^2.1.1",
-    "lucide-react": "^0.236.0",
+    "lucide-react": "^0.252.0",
     "qrcode.react": "^3.1.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
