+++ frontend/package.json.md5
@@ -1,1 +1,1 @@
-8d009b842dcc1048f3b2e25d59e7993a
+d6a6be868fff6e891ca25c1fc3c4d1b8
